Refactor app-schema & gt-complex to avoid wfs-ng->app-schema dependency
