+++ modules/extension/app-schema/app-schema/src/main/java/org/geotools/data/complex/NestedAttributeMapping.java
@@ -29,11 +29,11 @@ import net.opengis.wfs20.ResolveValueType;
 
 import org.geotools.data.FeatureSource;
 import org.geotools.data.Query;
+import org.geotools.data.complex.config.Types;
 import org.geotools.data.complex.filter.XPathUtil.StepList;
 import org.geotools.factory.Hints;
 import org.geotools.feature.FeatureCollection;
 import org.geotools.feature.FeatureIterator;
-import org.geotools.feature.Types;
 import org.geotools.filter.FilterFactoryImplNamespaceAware;
 import org.geotools.util.Converters;
 import org.opengis.feature.Feature;
