+++ modules/extension/app-schema/app-schema/src/main/java/org/geotools/data/complex/XmlMappingFeatureIterator.java
@@ -28,12 +28,12 @@ import java.util.logging.Level;
 import org.apache.commons.jxpath.JXPathException;
 import org.geotools.data.Query;
 import org.geotools.data.complex.PathAttributeList.Pair;
+import org.geotools.data.complex.config.Types;
 import org.geotools.data.complex.filter.XPath.*;
 import org.geotools.data.complex.filter.XPathUtil.*;
 import org.geotools.data.complex.xml.*;
-import org.geotools.feature.AttributeBuilder;
+import org.geotools.feature.AppSchemaAttributeBuilder;
 import org.geotools.feature.FeatureIterator;
-import org.geotools.feature.Types;
 import org.geotools.feature.type.ComplexTypeImpl;
 import org.geotools.filter.LiteralExpressionImpl;
 import org.geotools.util.XmlXpathUtilites;
@@ -173,7 +173,7 @@ public class XmlMappingFeatureIterator extends DataAccessMappingFeatureIterator 
     protected Feature populateFeatureData() throws IOException {
 
         final AttributeDescriptor targetNode = mapping.getTargetFeature();
-        AttributeBuilder builder = new AttributeBuilder(attf);
+        AppSchemaAttributeBuilder builder = new AppSchemaAttributeBuilder(attf);
         builder.setDescriptor(targetNode);
         Feature target = (Feature) builder.build(extractIdForAttribute(mapping
                 .getFeatureIdExpression(), null));
