+++ modules/extension/app-schema/app-schema/src/test/java/org/geotools/data/complex/GeologicUnitTest.java
@@ -20,23 +20,25 @@ package org.geotools.data.complex;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
+
 import java.io.IOException;
 import java.net.URL;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.Map;
 import java.util.Set;
+
 import org.geotools.data.DataAccess;
 import org.geotools.data.DataAccessFinder;
 import org.geotools.data.FeatureSource;
 import org.geotools.data.complex.config.AppSchemaDataAccessConfigurator;
 import org.geotools.data.complex.config.AppSchemaDataAccessDTO;
-import org.geotools.data.complex.config.EmfAppSchemaReader;
+import org.geotools.data.complex.config.EmfComplexFeatureReader;
 import org.geotools.data.complex.config.AppSchemaFeatureTypeRegistry;
+import org.geotools.data.complex.config.Types;
 import org.geotools.data.complex.config.XMLConfigDigester;
 import org.geotools.feature.FeatureCollection;
 import org.geotools.feature.FeatureIterator;
-import org.geotools.feature.Types;
 import org.geotools.test.AppSchemaTestSupport;
 import org.geotools.xml.SchemaIndex;
 import org.junit.BeforeClass;
@@ -63,7 +65,7 @@ public class GeologicUnitTest extends AppSchemaTestSupport {
 
     private static final String schemaBase = "/test-data/";
 
-    private static EmfAppSchemaReader reader;
+    private static EmfComplexFeatureReader reader;
 
     /**
      * Set up the reader
@@ -73,7 +75,7 @@ public class GeologicUnitTest extends AppSchemaTestSupport {
      */
     @BeforeClass
     public static void oneTimeSetUp() throws Exception {
-        reader = EmfAppSchemaReader.newInstance();
+        reader = EmfComplexFeatureReader.newInstance();
     }
 
     /**
