+++ modules/unsupported/app-schema/webservice/src/test/java/org/geotools/data/complex/XmlDataStoreTest.java
@@ -38,6 +38,7 @@ import org.geotools.data.FeatureSource;
 import org.geotools.data.Query;
 import org.geotools.data.complex.config.AppSchemaDataAccessConfigurator;
 import org.geotools.data.complex.config.AppSchemaDataAccessDTO;
+import org.geotools.data.complex.config.Types;
 import org.geotools.data.complex.config.XMLConfigDigester;
 import org.geotools.data.complex.xml.XmlResponse;
 import org.geotools.data.ws.WSDataStoreFactory;
@@ -47,7 +48,6 @@ import org.geotools.feature.ComplexAttributeImpl;
 import org.geotools.feature.FeatureCollection;
 import org.geotools.feature.FeatureIterator;
 import org.geotools.feature.NameImpl;
-import org.geotools.feature.Types;
 import org.geotools.filter.FilterFactoryImplNamespaceAware;
 import org.geotools.gml3.bindings.GML3EncodingUtils;
 import org.geotools.util.Converters;
